requirements.txt
manim>=0.18.0
ttkbootstrap
Pillow 